Redid Task Id Feature
Made task id's not only present in the key attribute for React, but also a key in the array of task items. When the user clicks the x button, the html key attribute is targeted, and any item with that same id in the items array is filtered out. Also changed some stylings with the submit buttons, glow shadows, and input box.